+++ pydashlite/collections/__init__.py
@@ -6,5 +6,6 @@ from .chunk import chunk
 from .set_ import set_
 from .merge import merge
 from .pick import pick, pickDeep
+from .unset import unset
 
-__all__ = ['get', 'groupBy', 'head', 'at', 'chunk', 'set_', 'merge', 'pick', 'pickDeep']
+__all__ = ['get', 'groupBy', 'head', 'at', 'chunk', 'set_', 'merge', 'pick', 'pickDeep', 'unset']